+++ offchain/crates/solana-cli/src/command/shreds/mod.rs
@@ -2,6 +2,7 @@ pub mod list;
 pub mod pay;
 pub mod payments;
 pub mod price;
+pub mod validator_client_rewards;
 pub mod withdraw;
 
 use anyhow::{Result, bail};
@@ -45,6 +46,9 @@ pub enum ShredsSubcommand {
     Payments(payments::PaymentsCommand),
     /// Show current device pricing.
     Price(price::PriceCommand),
+    /// Set the rewards proportion for a validator client.
+    #[command(hide = true)]
+    ValidatorClientRewards(validator_client_rewards::ValidatorClientRewardsCommand),
 }
 
 impl ShredsSubcommand {
@@ -55,6 +59,7 @@ impl ShredsSubcommand {
             Self::List(command) => command.try_into_execute(dz_ledger_url).await,
             Self::Payments(command) => command.try_into_execute(dz_ledger_url).await,
             Self::Price(command) => command.try_into_execute(dz_ledger_url).await,
+            Self::ValidatorClientRewards(command) => command.try_into_execute().await,
         }
     }
 }