+++ offchain/crates/solana-sdk/src/shred_subscription/state.rs
@@ -10,6 +10,7 @@ pub const CLIENT_SEAT_SEED_PREFIX: &[u8] = b"client_seat";
 pub const METRO_HISTORY_SEED_PREFIX: &[u8] = b"metro_history";
 pub const TOKEN_PDA_SEED_PREFIX: &[u8] = b"token";
 pub const PAYMENT_ESCROW_SEED_PREFIX: &[u8] = b"payment_escrow";
+pub const VALIDATOR_CLIENT_REWARDS_SEED_PREFIX: &[u8] = b"validator_client_rewards";
 pub const INSTANT_ALLOCATION_REQUEST_SEED_PREFIX: &[u8] = b"instant_seat_allocation_request";
 pub const WITHDRAW_SEAT_REQUEST_SEED_PREFIX: &[u8] = b"withdraw_seat_request";
 
@@ -58,6 +59,16 @@ pub fn find_token_pda_address(token_owner_key: &Pubkey, mint_key: &Pubkey) -> (P
             TOKEN_PDA_SEED_PREFIX,
             token_owner_key.as_ref(),
             mint_key.as_ref(),
+        ],
+        &crate::shred_subscription::ID,
+    )
+}
+
+pub fn find_validator_client_rewards_address(client_id: u16) -> (Pubkey, u8) {
+    Pubkey::find_program_address(
+        &[
+            VALIDATOR_CLIENT_REWARDS_SEED_PREFIX,
+            &client_id.to_le_bytes(),
         ],
         &crate::shred_subscription::ID,
     )